add contact component
